simulating-quantum-systems/requirements.txt
qiskit==0.19.2
numpy==1.20.3
matplotlib==3.2.2
seaborn==0.11.2
pandas==1.3.4
plotly==5.6.0
streamlit==1.7.0
streamlit-option-menu==0.3.2
hydralit-components==1.0.9
psycopg2-binary==2.9.3
h5py==3.2.1
plotly-express==0.4.1
streamlit-aggrid==0.2.3
